WES Kubernetes deployment
Add a new deployment directory that contains YAML files for deploying
WES on Kubernetes/OpenShift. Separate deployments are created for the
WES Flask server and the Celery worker. These can then be independently
scaled up or down. A shared volume is created for the Flask app and the
Celery worker to exchange data between each other.

Add a Route object under deploymen/ingress for creating a Route to
expose WES in OpenShift. The Route is set to use TLS and to redirect
HTTP traffic to HTTPS. It is assumed that a default wildcard certificate
is available.

Add documentation for setting up WES in Kubernetes under
deployment/README.md and reference this file in the main README.md file.