MongoDB and RabbitMQ env variables
In order to make the deployment compatible with OpenShift/K8S,
settings for MongoDB and RabbitMQ have been introduced as
environment variables.

The environment variables are:
* MONGO_HOST
* MONGO_PORT
* MONGO_DBNAME
* MONGO_USERNAME
* MONGO_PASSWORD
* RABBIT_HOST
* RABBIT_PORT

If the aforementioned environment variables are not set, the
settings from wes_elixir/config/app_config.yaml will be sourced.
This makes the deployment backwards compatible with
docker-compose.

The Dockerfile for the Flask server and Celerty app are still in a
common file. The "ENTRYPOINTS" for Openshift/K8S will be
specified in the deployment YAMLs.