Document RabbitMQ template used for testing
Add a link in the deployment documentation to the RabbitMQ cluster
template that was used while developing the WES deployment YAML files.
This template only works in OpenShift, but it is documented here so that
all components of the current WES deployment are documented.

The deployment documentation is getting long enough for a table of
contents to be useful, so add one.